accounts_service/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
docker==7.1.0
psycopg2==2.9.9
requests==2.32.3
sqlparse==0.5.1
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==5.1.1
django-admin==2.0.2
django-excel-response2==3.0.6
django-six==1.0.5
excel-base==1.0.4
idna==3.10
isoweek==1.3.3
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
requests-oauthlib==2.0.0
screen==1.0.1
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
TimeConvert==3.0.13
tzlocal==5.2
urllib3==2.2.3
xlwt==1.3.0
django-allauth==0.52.0  
Pillow==10.4.0
gunicorn==23.0.0
packaging==24.1
pyotp
qrcode
python-dotenv